Remove useless casts

+++ AetherLogger.Models/DegMinSec.cs
@@ -8,8 +8,8 @@ public struct DegMinSec
 
     public DegMinSec(double value)
     {
-        Degrees = (double)(int)value;
-        Minutes = (double)(int)((value - Degrees) * 60);
+        Degrees = (int)value;
+        Minutes = (int)((value - Degrees) * 60);
         Seconds = (value - Degrees - Minutes / 60.0) * 3600.0;
     }
